wip: replaced svg to png and working responsive
