+++ package.json
@@ -18,18 +18,20 @@
     "core-js": "2.6.3",
     "emotion": "^10.0.7",
     "emotion-server": "^10.0.7",
+    "facepaint": "^1.2.1",
     "gatsby": "2.0.118",
     "gatsby-image": "^2.0.29",
     "gatsby-plugin-catch-links": "^2.0.10",
     "gatsby-plugin-emotion": "4.0.1",
     "gatsby-plugin-manifest": "2.0.17",
+    "gatsby-plugin-material-ui": "^1.2.4",
     "gatsby-plugin-offline": "2.0.23",
     "gatsby-plugin-react-helmet": "3.0.6",
     "gatsby-plugin-sharp": "^2.0.20",
     "gatsby-plugin-typescript": "2.0.7",
     "gatsby-remark-component": "^1.1.3",
     "gatsby-remark-copy-linked-files": "^2.0.9",
-    "gatsby-remark-images": "^3.0.3",
+    "gatsby-remark-images": "3.0.4",
     "gatsby-remark-prismjs": "^3.2.4",
     "gatsby-source-filesystem": "2.0.20",
     "gatsby-transformer-remark": "2.2.4",
