+++ src/pages/404.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Typography from '@material-ui/core/Typography';
 
-import Layout from '../components/Layout';
+import Layout from '../layouts/Layout';
 
 const NotFoundPage = () => (
   <Layout>
