+++ src/pages/book-api.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { StaticQuery, graphql } from 'gatsby';
 import SwaggerUI from '../components/SwaggerUI';
-import Layout from '../components/Layout';
+import Layout from '../layouts/Layout';
 
 const BookApiPage = () => (
   <StaticQuery
